backend/backend.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import os
from PIL import Image
from textblob import TextBlob
import language_tool_python
import requests
import pandas as pd
import random
import eng_to_ipa as ipa
from azure.cognitiveservices.vision.computervision import ComputerVisionClient
from azure.cognitiveservices.vision.computervision.models import OperationStatusCodes
from msrest.authentication import CognitiveServicesCredentials
import time
from abydos.phonetic import Soundex, Metaphone, Caverphone, NYSIIS

app = Flask(__name__)
CORS(app)  # Enable CORS for React frontend

# Initialize your existing functions and variables here
subscription_key_imagetotext = "1780f5636509411da43040b70b5d2e22"
endpoint_imagetotext = "https://prana-------------v.cognitiveservices.azure.com/"
api_key_textcorrection = "7aba4995897b4dcaa86c34ddb82a1ecf"
endpoint_textcorrection = "https://api.bing.microsoft.com/v7.0/SpellCheck"

try:
    computervision_client = ComputerVisionClient(
        endpoint_imagetotext, CognitiveServicesCredentials(subscription_key_imagetotext))
    my_tool = language_tool_python.LanguageTool('en-US')
except Exception as e:
    print(f"Warning: Could not initialize services: {e}")
    computervision_client = None
    my_tool = None

# Copy all your existing functions here
def levenshtein(s1, s2):
    if len(s1) < len(s2):
        return levenshtein(s2, s1)
    if len(s2) == 0:
        return len(s1)
    previous_row = range(len(s2) + 1)
    for i, c1 in enumerate(s1):
        current_row = [i + 1]
        for j, c2 in enumerate(s2):
            insertions = previous_row[j + 1] + 1
            deletions = current_row[j] + 1
            substitutions = previous_row[j] + (c1 != c2)
            current_row.append(min(insertions, deletions, substitutions))
        previous_row = current_row
    return previous_row[-1]

def image_to_text(image_path):
    if computervision_client is None:
        return "Sample text for testing purposes"
    
    try:
        read_image = open(image_path, "rb")
        read_response = computervision_client.read_in_stream(read_image, raw=True)
        read_operation_location = read_response.headers["Operation-Location"]
        operation_id = read_operation_location.split("/")[-1]

        while True:
            read_result = computervision_client.get_read_result(operation_id)
            if read_result.status.lower() not in ['notstarted', 'running']:
                break
            time.sleep(5)

        text = []
        if read_result.status == OperationStatusCodes.succeeded:
            for text_result in read_result.analyze_result.read_results:
                for line in text_result.lines:
                    text.append(line.text)

        return " ".join(text)
    except Exception as e:
        print(f"Error in image_to_text: {e}")
        return "Sample text for testing purposes"

def spelling_accuracy(extracted_text):
    try:
        spell_corrected = TextBlob(extracted_text).correct()
        return ((len(extracted_text) - (levenshtein(extracted_text, str(spell_corrected))))/(len(extracted_text)+1))*100
    except Exception as e:
        return 85.0

def gramatical_accuracy(extracted_text):
    try:
        spell_corrected = TextBlob(extracted_text).correct()
        if my_tool is not None:
            correct_text = my_tool.correct(str(spell_corrected))
        else:
            correct_text = str(spell_corrected)
        extracted_text_set = set(str(spell_corrected).split(" "))
        correct_text_set = set(correct_text.split(" "))
        n = max(len(extracted_text_set - correct_text_set),
                len(correct_text_set - extracted_text_set))
        return ((len(str(spell_corrected)) - n)/(len(str(spell_corrected))+1))*100
    except Exception as e:
        return 80.0

def percentage_of_corrections(extracted_text):
    try:
        data = {'text': extracted_text}
        params = {'mkt': 'en-us', 'mode': 'proof'}
        headers = {
            'Content-Type': 'application/x-www-form-urlencoded',
            'Ocp-Apim-Subscription-Key': api_key_textcorrection,
        }
        response = requests.post(endpoint_textcorrection, headers=headers, params=params, data=data)
        json_response = response.json()
        
        if 'flaggedTokens' in json_response:
            return len(json_response['flaggedTokens'])/len(extracted_text.split(" "))*100
        else:
            return 5.0
    except Exception as e:
        return min(20.0, len(extracted_text.split()) * 2.0)

def percentage_of_phonetic_accuraccy(extracted_text: str):
    # Copy your existing phonetic accuracy function here
    soundex = Soundex()
    metaphone = Metaphone()
    caverphone = Caverphone()
    nysiis = NYSIIS()
    spell_corrected = TextBlob(extracted_text).correct()

    # ... rest of your phonetic accuracy logic
    return 85.0  # placeholder

def score(input_features):
    # Copy your existing ML prediction logic here
    if input_features[0] <= 96.40350723266602:
        var0 = [0.0, 1.0]
    else:
        if input_features[1] <= 99.1046028137207:
            var0 = [0.0, 1.0]
        else:
            if input_features[2] <= 2.408450722694397:
                if input_features[2] <= 1.7936508059501648:
                    var0 = [1.0, 0.0]
                else:
                    var0 = [0.0, 1.0]
            else:
                var0 = [1.0, 0.0]
    return var0

# API Routes
@app.route('/api/analyze-image', methods=['POST'])
def analyze_image():
    try:
        if 'file' not in request.files:
            return jsonify({'error': 'No file uploaded'}), 400
        
        file = request.files['file']
        if file.filename == '':
            return jsonify({'error': 'No file selected'}), 400
        
        # Save uploaded file temporarily
        temp_path = f"temp_{int(time.time())}.jpg"
        file.save(temp_path)
        
        # Extract text and analyze
        extracted_text = image_to_text(temp_path)
        
        # Calculate metrics
        spelling_acc = spelling_accuracy(extracted_text)
        grammatical_acc = gramatical_accuracy(extracted_text)
        corrections_pct = percentage_of_corrections(extracted_text)
        phonetic_acc = percentage_of_phonetic_accuraccy(extracted_text)
        
        # Get prediction
        features = [spelling_acc, grammatical_acc, corrections_pct, phonetic_acc]
        prediction = score(features)
        
        # Clean up temp file
        os.remove(temp_path)
        
        return jsonify({
            'extracted_text': extracted_text,
            'spelling_accuracy': spelling_acc,
            'grammatical_accuracy': grammatical_acc,
            'percentage_of_corrections': corrections_pct,
            'phonetic_accuracy': phonetic_acc,
            'result': prediction[0] == 1.0
        })
    
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/get-words', methods=['GET'])
def get_words():
    try:
        level = int(request.args.get('level', 1))
        
        if level == 1:
            csv_file = 'data/intermediate_voc.csv'
        elif level == 2:
            csv_file = 'data/elementary_voc.csv'
        else:
            return jsonify({'error': 'Invalid level'}), 400
            
        if os.path.exists(csv_file):
            voc = pd.read_csv(csv_file)
            arr = voc.squeeze().to_numpy()
            selected_list = random.sample(list(arr), min(10, len(arr)))
            return jsonify({'words': selected_list})
        else:
            # Fallback words
            fallback_words = ["apple", "banana", "cat", "dog", "elephant", "fish", "giraffe", "house", "ice", "jacket"]
            return jsonify({'words': fallback_words})
            
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/check-pronunciation', methods=['POST'])
def check_pronunciation():
    try:
        data = request.get_json()
        original = data.get('original', '')
        pronounced = data.get('pronounced', '')
        
        # Convert to IPA
        original_ipa = ipa.convert(original)
        pronounced_ipa = ipa.convert(pronounced)
        
        # Calculate inaccuracy
        inaccuracy = levenshtein(original_ipa, pronounced_ipa) / len(original_ipa) if len(original_ipa) > 0 else 0
        
        return jsonify({
            'original_ipa': original_ipa,
            'pronounced_ipa': pronounced_ipa,
            'inaccuracy': inaccuracy
        })
    
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/check-dictation', methods=['POST'])
def check_dictation():
    try:
        data = request.get_json()
        words = data.get('words', [])
        user_input = data.get('user_input', [])
        
        accuracy = []
        for i, (word, input_word) in enumerate(zip(words, user_input)):
            if len(word) > 0:
                acc = 1 - (levenshtein(word.lower(), input_word.lower()) / len(word))
                accuracy.append(max(0, acc))
            else:
                accuracy.append(0)
        
        overall_accuracy = sum(accuracy) / len(accuracy) if len(accuracy) > 0 else 0
        
        return jsonify({
            'accuracy': accuracy,
            'overall_accuracy': overall_accuracy
        })
    
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)