backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==8.3.1
textblob==0.15.3
language-tool-python==2.7.1
requests==2.26.0
pandas==1.3.3
SpeechRecognition==3.8.1
pyttsx3==2.90
eng-to-ipa==0.0.2
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.6.21
abydos==0.5.0